move photoemission feature in CameraInstrument
