+++ laserstudio/instruments/camera_raptor.py
@@ -187,30 +187,14 @@ class CameraRaptorInstrument(CameraUSBInstrument):
 
         self.last_frame_number = 0
 
-        self._last_frame_accumulator: numpy.ndarray = numpy.zeros(
-            self.width * self.height, dtype=numpy.uint64
-        )
-        self._image_averaging = 1
-        self._last_frames: list[numpy.ndarray] = [
-            numpy.zeros(self.width * self.height, dtype=numpy.uint16)
-        ] * self._image_averaging
-
-        self.reference_image_accumulator: Optional[numpy.ndarray] = None
-
         self.manufacturers_data = None
 
         # Objective on this camera is 10 by default
         objective = cast(float, config.get("objective", 10.0))
         self.select_objective(objective)
 
-        # By default, show "negative" activity, eg the pixels that are less than the reference image
-        self.show_negative_values = True
-
-        # Window averaging makes to store all averaged image to make a rotating average
-        # When the number of images to average is hit, and a new frame is retrieved,
-        # the oldest one is removed from the accumulator and the new one is added.
-        self.windowed_averaging = True
-        self.number_of_averaged_images = 0
+        # Reference image feature
+        self.reference_image_accumulator: Optional[numpy.ndarray] = None
 
     def query_command(
         self,
@@ -269,164 +253,6 @@ class CameraRaptorInstrument(CameraUSBInstrument):
         )
         value = self.get_value_at_address(address, expected_bytes)
         return int.from_bytes(value, "big")
-
-    @property
-    def image_averaging(self) -> int:
-        """
-        Returns the number of images that must be averaged.
-        """
-        return self._image_averaging
-
-    @image_averaging.setter
-    def image_averaging(self, value: int):
-        """
-        Sets the number of images that must be averaged.
-        """
-        if self.windowed_averaging:
-            if len(self._last_frames) > value:
-                # Drop the oldest frames (at the begining of the array)
-                to_substract = self._last_frames[: len(self._last_frames) - value]
-                self._last_frames = self._last_frames[len(self._last_frames) - value :]
-                self._last_frame_accumulator -= sum(to_substract)
-        else:
-            self.clear_averaged_images()
-
-        self._image_averaging = value
-
-    def clear_averaged_images(self):
-        """
-        Clears the list of averaged images.
-        """
-        self._last_frames = []
-        self._last_frame_accumulator = numpy.zeros(
-            self.width * self.height, dtype=numpy.uint64
-        )
-        self.number_of_averaged_images = 0
-
-    # @property
-    # def last_frame(self) -> numpy.ndarray:
-    #     """
-    #     Return the frame that should be analysed by histogram computation
-    #     """
-    #     return (
-    #         ((self._last_frame_accumulator * 4) / self.average_count)
-    #         .clip(
-    #             min=numpy.iinfo(numpy.uint16).min,
-    #             max=numpy.iinfo(numpy.uint16).max,
-    #         )
-    #         .astype(numpy.uint16)
-    #     )
-
-    def accumulate_frame(self, new_frame: numpy.ndarray):
-        """
-        Accumulates the given frame and removes the oldest one
-          if windowed averaging is active.
-        """
-        if not self.windowed_averaging:
-            if self.number_of_averaged_images == self._image_averaging:
-                # Discarding the new frame from accumulation
-                return
-        if self._image_averaging == self.number_of_averaged_images:
-            # The list is full, we can remove the oldest frame
-            self._last_frame_accumulator -= self._last_frames.pop(0)
-            self.number_of_averaged_images -= 1
-        # Add the new frame in the accumulator and the list
-        self._last_frames.append(new_frame)
-        self._last_frame_accumulator += new_frame
-        self.number_of_averaged_images += 1
-
-    @property
-    def average_count(self) -> int:
-        """
-        Returns the number of images that have been averaged.
-        """
-        return self.number_of_averaged_images
-
-    def get_last_image(
-        self,
-    ) -> tuple[int, int, Literal["L", "I;16", "RGB"], Optional[bytes]]:
-        ret, frame = self.vc.read()
-        if not ret or frame is None:
-            return self.width, self.height, "RGB", None
-        assert type(frame) is numpy.ndarray
-        # Each value is repeated three times...
-        frame = frame[:, :, :1].copy()
-        # Flatten the array
-        frame = numpy.reshape(frame, (-1,))
-        # Frame number is present in the data
-        number = frame[0:8]
-        self.last_frame_number = number.view(numpy.uint32)[0]
-        # Remove the 8 first bytes (containing the frame number)
-        frame = frame[8:]
-        # Interpret as 16 bits array
-        frame = frame.view(numpy.uint16).copy()
-        # Add 0s to the end to compensate the values that were removed for frame number
-        frame = numpy.resize(frame, self.width * self.height)
-        # Put the frame in the accumulator
-        self.accumulate_frame(frame)
-        # Apply the subtraction of reference image
-        pos, neg = self.substract_reference_image()
-
-        # Apply levels
-        pos = self.apply_levels(pos)
-        if neg is not None:
-            neg = self.apply_levels(neg)
-
-        frame = self.construct_display_image(pos, neg)
-        if frame.dtype == numpy.uint16:
-            frame_bytes = frame.byteswap().tobytes()
-            mode = "I;16"
-        else:
-            frame_bytes = frame.tobytes()
-            mode = "RGB"
-
-        return (
-            self.width,
-            self.height,
-            mode,
-            frame_bytes,
-        )
-
-    @property
-    def last_frame(self) -> numpy.ndarray:
-        """
-        Return the frame that should be analysed by histogram computation
-        """
-        pos = self._last_pos
-        neg = self._last_neg
-        return self.construct_display_image(pos, neg)
-
-    def construct_display_image(self, pos: numpy.ndarray, neg) -> numpy.ndarray:
-        """
-        Construct the display image from the positive and negative images.
-        """
-        pos_16 = (
-            (pos * (4.0 / self.average_count))
-            .clip(
-                min=numpy.iinfo(numpy.uint16).min,
-                max=numpy.iinfo(numpy.uint16).max,
-            )
-            .astype(numpy.uint16)
-        )
-        if neg is None:
-            return pos_16
-
-        neg_16 = (
-            (neg * (4.0 / self.average_count))
-            .clip(
-                min=numpy.iinfo(numpy.uint16).min,
-                max=numpy.iinfo(numpy.uint16).max,
-            )
-            .astype(numpy.uint16)
-        )
-        zer_16 = numpy.zeros((self.height, self.width, 1), dtype=numpy.uint16)
-        pos_16 = pos_16.reshape(zer_16.shape)
-        neg_16 = neg_16.reshape(zer_16.shape)
-        stacked = numpy.stack([neg_16, pos_16, zer_16], axis=2)
-
-        stacked = stacked >> 4
-        # 8 bits per pixel
-        return stacked.astype(numpy.uint8).reshape(self.height * self.width * 3)
 
     def get_micro_version(self) -> tuple[int, int]:
         """
@@ -660,34 +486,47 @@ class CameraRaptorInstrument(CameraUSBInstrument):
         self.temperature_changed.emit(temperature)
         return temperature
 
-    def take_reference_image(self, do_take: bool):
-        """
-        Take a reference image to substract from the next frames.
-        """
-        if do_take:
-            self.reference_image_accumulator = self._last_frame_accumulator.copy()
-        else:
-            self.reference_image_accumulator = None
-
-    def substract_reference_image(
+    def get_last_image(
         self,
-    ) -> tuple[numpy.ndarray, Optional[numpy.ndarray]]:
-        """Substract the reference_image_accumulator from the current accumulator"""
-        if self.reference_image_accumulator is None:
-            """The reference image is not yet defined"""
-            self._last_pos = self._last_frame_accumulator
-            self._last_neg = None
+    ) -> tuple[int, int, Literal["L", "I;16", "RGB"], Optional[bytes]]:
+        ret, frame = self.vc.read()
+        if not ret or frame is None:
+            return self.width, self.height, "RGB", None
+        assert type(frame) is numpy.ndarray
+        # Each value is repeated three times...
+        frame = frame[:, :, :1].copy()
+        # Flatten the array
+        frame = numpy.reshape(frame, (-1,))
+        # Frame number is present in the data
+        number = frame[0:8]
+        self.last_frame_number = number.view(numpy.uint32)[0]
+        # Remove the 8 first bytes (containing the frame number)
+        frame = frame[8:]
+        # Interpret as 16 bits array
+        frame = frame.view(numpy.uint16).copy()
+        # Add 0s to the end to compensate the values that were removed for frame number
+        frame = numpy.resize(frame, self.width * self.height)
+        # Put the frame in the accumulator
+        self.accumulate_frame(frame)
+        # Apply the subtraction of reference image
+        pos, neg = self.substract_reference_image()
+
+        # Apply levels
+        pos = self.apply_levels(pos)
+        if neg is not None:
+            neg = self.apply_levels(neg)
+
+        frame = self.construct_display_image(pos, neg)
+        if frame.dtype == numpy.uint16:
+            frame_bytes = frame.byteswap().tobytes()
+            mode = "I;16"
         else:
-            self._last_pos = (
-                (self._last_frame_accumulator - self.reference_image_accumulator)
-                .astype(numpy.int64)
-                .clip(0)
-                .astype(numpy.uint64)
-            )
-            self._last_neg = (
-                (self.reference_image_accumulator - self._last_frame_accumulator)
-                .astype(numpy.int64)
-                .clip(0)
-                .astype(numpy.uint64)
-            )
-        return self._last_pos, self._last_neg
+            frame_bytes = frame.tobytes()
+            mode = "RGB"
+
+        return (
+            self.width,
+            self.height,
+            mode,
+            frame_bytes,
+        )
